fixed warnings

+++ applications/mp4box/fileimport.c
@@ -80,7 +80,7 @@ GF_Err set_file_udta(GF_ISOFile *dest, u32 tracknum, u32 udta_type, char *src, B
 #endif
 	if (is_string) {
 		data = (u8 *) src;
-		size = strlen(src)+1;
+		size = (u32) strlen(src)+1;
 		is_box_array = 0;
 	} else {
 		GF_Err e = gf_file_load_data(src, (u8 **) &data, &size);
